Moved syncing to another thread and incorporated current status into UI
