+++ src/main/java/coreapp/DriveInitializer.java
@@ -1,0 +1,114 @@
+package coreapp;
+
+import com.google.api.client.auth.oauth2.Credential;
+import com.google.api.client.extensions.java6.auth.oauth2.AuthorizationCodeInstalledApp;
+import com.google.api.client.extensions.jetty.auth.oauth2.LocalServerReceiver;
+import com.google.api.client.googleapis.auth.oauth2.GoogleAuthorizationCodeFlow;
+import com.google.api.client.googleapis.auth.oauth2.GoogleClientSecrets;
+import com.google.api.client.googleapis.javanet.GoogleNetHttpTransport;
+import com.google.api.client.http.HttpTransport;
+import com.google.api.client.json.JsonFactory;
+import com.google.api.client.json.jackson2.JacksonFactory;
+import com.google.api.client.util.store.FileDataStoreFactory;
+import com.google.api.services.drive.Drive;
+import com.google.api.services.drive.DriveScopes;
+
+import java.io.IOException;
+import java.io.InputStream;
+import java.io.InputStreamReader;
+import java.util.Arrays;
+import java.util.List;
+
+public class DriveInitializer {
+    /** Application name. */
+    private static final String APPLICATION_NAME =
+            "DriveSync";
+
+    /** Directory to store user credentials for this application. */
+    private static final java.io.File DATA_STORE_DIR = new java.io.File(
+            System.getProperty("user.home"), ".credentials/drive-java-drivesync");
+
+    /** Global instance of the {@link FileDataStoreFactory}. */
+    private static FileDataStoreFactory DATA_STORE_FACTORY;
+
+    /** Global instance of the JSON factory. */
+    private static final JsonFactory JSON_FACTORY =
+            JacksonFactory.getDefaultInstance();
+
+    /** Global instance of the HTTP transport. */
+    private static HttpTransport HTTP_TRANSPORT;
+
+    private static String ROOT_FOLDER_ID;
+
+    /** Global instance of the scopes required by this quickstart.
+     *
+     * If modifying these scopes, delete your previously saved credentials
+     * at ~/.credentials/drive-java-quickstart
+     */
+    private static final List<String> SCOPES =
+            Arrays.asList(DriveScopes.DRIVE);
+
+    static {
+        try {
+            HTTP_TRANSPORT = GoogleNetHttpTransport.newTrustedTransport();
+            DATA_STORE_FACTORY = new FileDataStoreFactory(DATA_STORE_DIR);
+        } catch (Throwable t) {
+            t.printStackTrace();
+            System.exit(1);
+        }
+    }
+
+    /**
+     * Creates an authorized Credential object.
+     * @return an authorized Credential object.
+     * @throws IOException
+     */
+    public static Credential authorize() throws IOException {
+        // Load client secrets.
+        int attempts = 0;
+
+        //Deletes saved credentials on fail
+        while (attempts < 3){
+            try{
+                InputStream in =
+                        SyncThread.class.getResourceAsStream("/client_secret.json");
+                GoogleClientSecrets clientSecrets =
+                        GoogleClientSecrets.load(JSON_FACTORY, new InputStreamReader(in));
+
+                // Build flow and trigger user authorization request.
+                GoogleAuthorizationCodeFlow flow =
+                        new GoogleAuthorizationCodeFlow.Builder(
+                                HTTP_TRANSPORT, JSON_FACTORY, clientSecrets, SCOPES)
+                                .setDataStoreFactory(DATA_STORE_FACTORY)
+                                .setAccessType("offline")
+                                .build();
+                Credential credential = new AuthorizationCodeInstalledApp(
+                        flow, new LocalServerReceiver()).authorize("user");
+                System.out.println(
+                        "Credentials saved to " + DATA_STORE_DIR.getAbsolutePath());
+                return credential;
+            }catch (IOException e){
+                attempts++;
+                DATA_STORE_DIR.delete();
+                if(attempts >= 3){
+                    throw e;
+                }
+            }
+        }
+        //Never called
+        return null;
+    }
+
+    /**
+     * Build and return an authorized Drive client service.
+     * @return an authorized Drive client service
+     * @throws IOException
+     */
+    public static Drive getDriveService() throws IOException {
+        Credential credential = authorize();
+        return new Drive.Builder(
+                HTTP_TRANSPORT, JSON_FACTORY, credential)
+                .setApplicationName(APPLICATION_NAME)
+                .build();
+    }
+}
