+++ src/main/java/coreapp/FrontEndGUI.java
@@ -1,7 +1,5 @@
 package coreapp;
 
-import com.google.api.services.drive.model.File;
-
 import java.io.IOException;
 
 /*
@@ -14,14 +12,13 @@ import java.io.IOException;
  *
  * @author Jason
  */
-public class FrontEndGUI extends javax.swing.JFrame {
+public class FrontEndGUI extends javax.swing.JFrame implements StatusUpdateCallback{
 
     /**
      * Creates new form FrontEndGUI
      */
     public FrontEndGUI() {
         initComponents();
-        syncFunctions = new SyncFunctions();
         //<editor-fold defaultstate="collapsed" desc=" Look and feel setting code (optional) ">
         /* If Nimbus (introduced in Java SE 6) is not available, stay with the default look and feel.
          * For details see http://download.oracle.com/javase/tutorial/uiswing/lookandfeel/plaf.html 
@@ -44,8 +41,13 @@ public class FrontEndGUI extends javax.swing.JFrame {
         }
         //</editor-fold>
     }
-    private SyncFunctions syncFunctions;
 
+
+    @Override
+    public void updateStatus(String text){
+        System.out.println("Callbacked");
+        lblStatus.setText(text);
+    }
     /**
      * This method is called from within the constructor to initialize the form.
      * WARNING: Do NOT modify this code. The content of this method is always
@@ -134,17 +136,22 @@ public class FrontEndGUI extends javax.swing.JFrame {
 
     private void btnEnterActionPerformed(java.awt.event.ActionEvent evt) {//GEN-FIRST:event_btnEnterActionPerformed
 
+        SyncThread syncThread;
         FileUtilities fileUtilities = new FileUtilities();
         String path = txtPathfield.getText();
         try
         {
             if(!fileUtilities.isValidPath(path)){
-                throw new IOException("Pleas enter a valid path.");
+                throw new IOException("Please enter a valid path.");
             }
-            Constants.ROOT_FOLDER_ID = SyncFunctions.updateMainDriveFolder().getId();
+
+
             path = fileUtilities.ensureEndSlash(path);
-            SyncFunctions.uploadFolder(path, Constants.ROOT_FOLDER_ID);
-            lblStatus.setText("Success");
+
+            syncThread = new SyncThread(this, path, Constants.ROOT_FOLDER_ID);
+            Constants.ROOT_FOLDER_ID = syncThread.updateMainDriveFolder().getId();
+            syncThread.start();
+            //lblStatus.setText("Success");
         }
         catch (IOException e)
         {
