+++ src/main/java/coreapp/StatusUpdateCallback.java
@@ -1,0 +1,5 @@
+package coreapp;
+
+public interface StatusUpdateCallback {
+    void updateStatus(String text);
+}
